Fix to run slicefix after all images have been loaded.
Otherwise we might run slicefix after a base image and before a delta
which could cause a train wreck in the FS if slicefix changes allocate
or deallocate blocks.

Also, grudgingly update some of the shell syntax to use constructs
introduced in the early 1990s. Sigh...this script will no longer run
on a PDP-11.