Add support for disabling dhcp for an experiment/user/group
In support of the collaboration effort with CHPC, I've added a check
and actions for a new feature, ExternalNodeManagment.  When swapping in,
the swapper, group, and experiment are checked for the feature.  Since
there doesn't appear to be a way to ask for an experiment feature in the
parser, I'm punting for now and just letting the experiment inherit this
feature from the user and/or project/group.

The effect of this feature is that nodes will be omitted from the
dhcpd.conf file when allocated to a marked experiment, and put back
in when freed from that experiment.

We will also use this feature to force nodes to reboot even when they
are in PXEWAIT during the OS setup phase.  Ultimately the point is for
something else to take over the boot process and manage the experiment's
nodes for the duration of the experiment.