Add 'Expect' CLI wrapper module for Force10 switch platform.
Not hooked up or tested yet.  Curses be to Force10 Networks that this
module had to be written at all.  There is really no other sane way to
manipulate vlan membership for portchannels (LAGs).  Force10's lame
excuse for a switch-side scripting API just telnets back in via the
loopback and uses Expect itself.  It also lacks functionality that we
require.  Horrific.