Tweak jquery-ui. The new version is trying to be too helpful.
Because of load ordering issues, we want to include jquery itself as a normal script rather than inside of require. The new version of jquery-ui tries to be smart and uses require if require exists. This will either fail if jquery isn't at the expected position or it will create a duplicate version of jquery that will not have the same set of plugins. Either way, bad. Just cut that bit of code out.