FreeNAS 9 support, "improved" FreeNAS 8 support.
A clientside top-level "gmake freenas-tarball" will build everything and
construct an appropriate tarball. You must either build on FreeBSD 8.3 or
FreeBSD 9.2, depending on the version of FreeNAS you are targetting.

This cannot be done native on the FreeNAS box! In part because there is
no compiler there, but even if there was, the install target would wreak
havoc on a full root filesystem; it assumes it is working on a skeleton
FS with just the Emulab stuff in it.

Mostly this commit is grotesque Makefile hacking due to our tragic
client-side tmcc OS-specific directory structure. Hey, don't blame me!
It was, um...okay DO blame me...