Export methods that interface to backend createdataset and deletelease scripts. Also add a describe method.
In order to create a dataset, the caller must either be a local user, or
have an extra credential that grants the "blockstores" priv.

Like slivers, datasets are currently created in the holding project and
group corresponding the SA and subauthority.

Eventually need to add sharing control (say, to make a dataset globally
read only), but that does not exist in the backend yet.