The latest in logic to have findSpareDisks not use the system disk.
If an available partition device (aka, the 4th partition on the system disk)
represents less than 5% of the spare space we have found, ignore it.

This will allow us to continue to use the 4th partition on the system
disk of the d710s (450GB or so) and the second disk (250GB), but not use
the 2nd partition (3GB), which would make us thrash about on the system
disk even more than usual.

Mostly this is for the new HP server boxes, so it doesn't pick up the 10GB
left over on the (virtual) system disk when we have 21TB available on the
second (virtual) disk.

Another hack til blockstores rule the world...