Add gid to project_leases, and allow creation in subgroups.
This is not exposed to users, the main reason for this is so that the name
space for leases (datasets) is per-group instead of per-project. We need
this when creating datasets via the geni interface (backend to APT), since
all leases are created in the holding project. Without a subgroup, we would
run into name collisions on the backend. It also gives us finer access
permission control for the same reason.

Note that I yanked out the lease cache from Lease.pm (not worth the
trouble), and I expanded Lookup to allow for the usual variety of
possibilities that we allow in other Lookup methods.