Account for "out of service" nodes.
In our usage info, let's not count nodes in hwdown and hwbroken as
"in use", as it make our node utilization overly high (well, at least for
pc600s and pc850s!)

Also, a couple more hacks to try to work around inconsistencies in the
node_history data. We really just need to fix up the history records!