Generalize bootlog read-more-from-socket code.
Long ago Leigh added code in the bootlog command to account for the fact that
you might not get all the data associated with a TCP-based command in one
read from the socket. This case also showed up with the hostkeys command
with uploaded multiple pubkeys.

So I moved the code up to handle_requests so that for all TCP-based
commands, we read til we get EOF.