Pull in new comware (h3c) module, courtesy of Keith Sklower @ DETER.
For now I'm adding this as a separate file, wherein I have renamed the
base procurve module as "snmpit_hpupd" until we have more time to test the
integration.  I did a fairly quick-and-dirty port of this code to work
with modern Emulab snmpit.

The comware module will be used for switches with "comware" in their
'type' string.