Add slots to apt_instances and apt_instance_history, to store the params and rspec that are used. This is to support Parameterized Profiles, where the rspec is generated from a geni-lib script and user defined parameters.
Add manifest to apt_instance_history; I was not storing the manifest cause
I figured we could just ask the backend cluster for it. Well, that is kinda
silly, so lets store it locally. I'll have to go back and grab the
manifests for the current history entries.

Add isaptkey to user_pubkeys so that we can mark the key that comes in from
the instantiate page. This lets us keep the Emulab keys for a user
independent of the APT key, so that APT users only deal with a single key
on that path, without messing up their Emulab keys.