ZFS change; when ZFS_NOEXPORTS is on, we have to call exports_setup when a user logs in so that their home directory and projects are exported from ops. Otherwise things break.
Since we really do not want to do this too often, exports_setup is
exporting anyone who is logged in within the last week, and the web
interface is calling out to exports_setup only once a day for each user.

This can be "improved" but I am worried we are fighting a losing battle and
will eventually yank this code anyway.