Add -w "wait for mountd to finish" option.
The heuristic we use here is that after ssh'ing to ops to HUP mountd,
we try to stat /proj/emulab-ops. We keep trying for 10 seconds or til
it works. In the new race-free NFS world, this stat will block until
mountd is done and then return success. In the old racy NFS world, the
stat should fail until mountd is done.