Change tiplines urlstamp to be an expiration time for the urlhash.
Previously it was the creation stamp for the hash. By making it the
expiration time, we can do different times for different nodes.

Note that there is no serious compatibility issue with re-purposing
the DB field. It is almost always zero (since they are only valid
for 5 minutes) and if it isn't zero when the new code is installed,
the hash will just immediately become invalid. So what? Big deal!