Preliminary "golden image" support using thin volumes in LVM.
Disabled for now. This is a checkpoint. This version still downloads
the compressed image into a volume and imageunzips into another volume.
The difference is that only one client does the imageunzip and then
everyone makes a snapshot of that.

On to getting rid of the initial download of the compressed image...