MFS fixups.
Make sure we clear out any partial GPTs. On FreeBSD we just use
"gpart destroy" which will get rid of an MBR or GPT.

Tee the output of rc.frisbee into a file and upload that back to
boss in the event of a failure. We will see if this proves useful.