FreeBSD 11 (aka current) support.
Mostly I just wanted to test "the latest" FBSD kernel on Apt
so see if anything done recently would help with the hanging
console problem on the R320s (newer kernels actually only hang
for about 10 seconds and drop around 500 chars when they probe
the ethernet interfaces--this is better than it was).

While in there, I just went ahead and did the client side,
mostly ifdefs in the makefiles.