Update Force10 module to dynamically determine size of PortSet vectors.
It's clear from the various Force10 switches and firmware versions
that we've seen now that the vlan membership PortSet bit
vector size can vary widely.  This variation even happens between
different versions of firmware on the same switch!  So, now we
detect this dynamically when the backend force10 snmpit object is
instantiated (in readifIndex()).