Add sitevar to determine whether clients should use UDP or TCP for NFS.
Yes, out of the blue and off the wall. But I got tired of trying to
guess what we had Linux and FreeBSD use. I was surprised to discover
that we were using UDP on Linux (which caused Clemson CloudLab to fail
because they have jumbo frames enabled on their control net switches
but ops had the MTU set to 1500).

Anyway, here it is. The default setting is UDP for backward compat.
We should probably set it to TCP nowadays. There is also an 'osdefault'
setting which says use the default setting on the client OS.