Add "disable_if" node/node_type attribute that can be returned via loadinfo.
This can be set (e.g. "bge1") to tell slicefix to set the appropriate
hint in loader.conf to tell FreeBSD to not even probe the indicated device.
We use this to prevent FreeBSD from screwing up the LOM port we use for the
IPMI management interface on the Apt R320 nodes.