Working well enough that I can start doing some serious testing.
To that end, added a new "imagerezip" program which reads in an image
and writes it back out. Right now, this is just to test that reading
and writing images via the new library produces the same result as the
"classic" tools. But it might be useful in the future to repack images
with a different compression level or a different signature (hash function
or blocksize).

Also checked in a hack to imageunzip I did for Xing (-I option) to just
produce a stream of the uncompressed image data (it ignores skips, aka
it doesn't skip on the output device to put the data in the correct place).

Unfortunately, I cannot just compare raw images because even trying to
replicate the chunk packing strategy of the original code, the new library
packs a bit more into each chunk. Not sure why, but anyway, I have to
compare old/new images by dumping their contents and by checking the
generated signatures.