Add an object definition to GeniHRN. I am getting tired of parsing URNs and picking pieces out, so I added this:
  my $foo = GeniHRN->new("urn:publicid:IDN+emulab.net:testbed+slice+myexp3");

and provides:

  $foo->domain();
  $foo->type();
  $foo->id();
  $foo->urn();

and has a stringify method that returns the urn, so that existing code all
works.

The problem though is the XML::RPC and Frontier libraries provide no hooks
to catch this, and rather then using a stringification, they both convert
all blessed references into structs, and so anyplace that puts a urn into
something to go out on the wire, has to be changed to force to the string.

Damn, how disappointing! So all the code is here but basically disabled
until I find time to go through all the code.