Respect the uploader_path field in image_versions.
The "Emulab configuration" of the mserver will now use uploader_path
if set rather than path.

A slight semantic change goes along with this. If the uploader_path is
set, we assume that it is a temporary file to upload to and we (the mserver)
do not do the dance of uploading to our own temporary and moving it into place.