Directory based image paths.
Soon, we will have images with both full images and deltas, for the same
image version. To make this possible, the image path will now be a
directory instead of a file, and all of the versions (ndz,sig,sha1,delta)
files will reside in the directory.

A new config variable IMAGEDIRECTORIES turns this on, there is also a check
for the ImageDiretories feature. This is applied only when a brand new
image is created; a clone version of the image inherits the path it started
with. Yes, you can have a mix of directory based and file based image
descriptors.

When it is time to convert all images over, there is a script called
imagetodir that will go through all image descriptors, create the
directory, move/rename all the files, and update the descriptors.
Ultimately, we will not support file based image paths.

I also added versioning to the image metadata descriptors so that going
forward, old clients can handle a descriptor from a new server.