Bug Fix: make port lookup distinguish logical/physical wires.
The stuff in db/Port.pm (Port::LookupByIface) wasn't handling the MLE
case where there are two wires table entries for physical wires with a
logical wire implemented by them; for my experiment, when snmpit was
trying to create the physical vlans on the apcon, it would grab the
logical wire instead of the physical one as it populated the Port
objects, and that screwed vlan creation up (the member port for the
physical layer 1 vlan gets labeled with a switch WIRE_END instead of
node -- which is correct for the logical wire, but not the physical).
So, I changed the query to check that wires.logical ==
interfaces.logical ; see the code.  This is a limited fix, maybe --
maybe that method should really do more verification to make sure it's
grabbed the correct wires table entry -- but since we only have 2 layers
it's plenty good for now.