Realpath strikes again.
Perl realpath returns undef or '' (take your pick) if the path does not
exist, rendering it pretty much useless for our check. So we run realpath
on the directory part of the path (which should exist) and then do some
other paranoid checks on the filename part (no funky chars, cannot be a
symlink).