Add OpenFlow support to HP Comware snmpit module (snmpit_h3c).
This commit fills in all of the snmpit OF API calls for the Comware
module.  Note that Comware doesn't support OF "listener" ports, such as
are available on the Procurve/Provision switches.  Therefore, the
functions related to listener ports return an error.

It is assumed that when "disableOpenflow()" is called, that the caller
also intends to destroy the related OpenFlow instance on the switch.
There is no other call in the snmpit API for explicitly removing OF
instances.