Don't do the named_wait thing on 10.1.
I had added use of this feature to make startup pause until the name
server was responding to avoid transient failures in elabinelab with
an "inner control net". But this feature does not work with a firewall
as it tries to do the named wait before any firewall rules have been loaded.