Support newer HP 5400zl firmware versions in libosload_switch.
First, firmwares are bigger and take longer to write, so bump the TFTP
load timeout to 5 minutes instead of 2.

Second, newer HP firmwares won't let you load unsigned firmware images
into flash, unless you specify a special param in your flash command.
Unfortunately, older firmwares don't support that param.  So we just use
trial and error to figure out if a switch has it or not, and we always
are willing to flash unsigned firmwares... just have to trust, because
we have no way of knowing which version should be signed, or how to tell
which version a particular blob is.  At least not that I know of, not
that I looked hard.

Finally, our expect-based interface has to be a bit more flexible
with some of the regexps it looks for.