Tweak os_islocaldir for m400 admin MFS where nothing is local.
Because the m400 MFS is NFS-mounted our traditional check to see if
a user homedir was local or NFS-mounted didn't work right. Now a homedir
is local if either it is really local or it is NFS-mounted but on the
same FS as "/". And we only do that latter on an MFS right now.