Fix up status rspec instruction templatizing to handle multi-site.
The little magic that runs the rspec_tour:instructions text through the
template code didn't work for multi-site, because it only looked at the
first manifest to generate its template values.  Well, if we have
multiple manifests, then nodes (and thus hostnames) are split between
manifests, so we have to look at all the manifests before we run the
instruction text through the templatizer.

So, we do it the fast way, building up our template data over the passes
through the manifests, saving off the last parsed manifest each time,
and then templatizing the rspec_tour:instructions from the last xml
blob... each manifest has a copy of the instructions.