Change API interface to connect via the node IP rather than "localhost".
In the current FreeNAS, the only options for configuring the web server
are to listen on a specific IP or to listen on INADDR_ANY. If you do the
former, then you cannot connect via "localhost". If you do the latter,
then any experiment with an iSCSI blockstore could connect (via its
per-blockstore VLAN/subnet/IP). While authentication is required to
connect, the latter still makes me nervous.

Note that I didn't have to change the "Dave API" to not use localhost
as it appears to hook in below the web server.