Support changing console and a couple of other bug fixes.
Do what we do in FreeBSD slicefix to set the console correctly.
Not perfect, but it is what it is.

Fix long standing bug that left garbage at the end of grub.cfg.
We were fixing grub.cfg by opening it for read/write, reading in the
whole file, fixing it, seeking to zero, and writing it back out.
However, if our fixes resulted in a shorter file than before, the
excess bits would be left in the file. Instead, we now truncate after
seeking to zero and before re-writing.

Fix another issue with CentOS 7. Make sure we look for linux16 and
initrd16 as well as linux and initrd.